Move quinticPolynomial
