+++ src/lib7842/api/positioning/path/quinticSegment.cpp
@@ -1,13 +1,33 @@
 #include "quinticSegment.hpp"
-#include "lib7842/api/other/quinticPolynomial.hpp"
 
 namespace lib7842 {
+
+QuinticPolynomial::QuinticPolynomial(double istart, double istartSlope, double iend,
+                                     double iendSlope) :
+  start(istart), startSlope(istartSlope), end(iend), endSlope(iendSlope) {}
+
+double QuinticPolynomial::calculate(double t) const {
+  double u = end - start - startSlope;
+  double v = endSlope - startSlope;
+
+  double a3 = 10 * u - 4 * v;
+  double a4 = -15 * u + 7 * v;
+  double a5 = 6 * u - 3 * v;
+
+  double coeffs[] = {start, startSlope, 0, a3, a4, a5};
+
+  double xt = 0;
+  for (size_t power = 0; power < 6; power++) {
+    xt += coeffs[power] * std::pow(t, power);
+  }
+
+  return xt;
+}
 
 QuinticSegment::QuinticSegment(const DataState& istart, const DataState& iend) :
   start(istart), end(iend) {}
 
 SimplePath QuinticSegment::generate(int isteps) const {
-
   double startSlope = start.getData<double>("slope");
   double endSlope = end.getData<double>("slope");
